Fix libkv source out-of-sync with Godeps
Looks like the libkv version vendored in really not in
sync with the git hash value in Godeps.json. The commit
https://github.com/docker/libnetwork/commit/e2a63dff5a38a0f26f7f0c36356f3a6a28f4efeb
has just updated the Godeps.json without update the source.
Dnet in multi-host testing is broken due to this, while
docker mult-host functionality works because the correct
version of libkv has been vendored in docker/docker.

Signed-off-by: Jana Radhakrishnan <mrjana@docker.com>