reserve the default gateway IP address if sharing the same subnet with host
If the user adds the physical interface to a bridge and configures
Docker to use the same subnet as the host (start daemon with the -b
parameter), the IP address of the default gateway might be in the
IP pool of the allocator.
In such case we should exclude it from the IP pool.

Signed-off-by: Shijiang Wei <mountkin@gmail.com>