Don't broadcast creation/deletion events across cluster.
A user makes one of the following requests in a single host:

network create
network delete
service publish
service unpublish
After the request is successful on one host, the same request is replicated
in every host. The requests are also sent again when docker daemon is
restarted. This means that in a 1000 node system, a single user request
gets converted into 1000 requests on the driver's backend.

This commit prevents this broadcast from happening.

Addresses the following enhancement:
https://github.com/docker/libnetwork/issues/313

Signed-off-by: Gurucharan Shetty <gshetty@nicira.com>