Ensure Keys are watchable with etcd
etcd v2.0+ with API v2 cannot watch on a regular key, we must enforce
the 'IsDir' parameter while creating the key for it to be considered
as a directory.

This step is necessary unless 'libkv' includes and supports the new
etcd API v3 which removes the directory/key distinction to allow for
a better abstraction.

Fixes: #392
See: docker/libkv#20

Signed-off-by: Alexandre Beslic <abronan@docker.com>