Fix IT test breakages
With DNS resolution happening within the container namespace
test should not try to ping a DNS name and expect a packet
loss message. It will only show up as a DNS name resolution
failure. Changed the bridge internal test to test for a
well known IP address.

Also rearranged the overlay internal tests so that it gets
to run before the dnet container is removed which was
happening in previous tests.

Signed-off-by: Jana Radhakrishnan <mrjana@docker.com>