Read veth stats from host
There is no need to do a setns to the network namespace to collect the
stats of a veth.  You can read the host side of the veth pairs stats,
from the host, and all you have to do is swap the rx/tx because of what
side you are reading from.

This removes the need for libnetwork to do a `setns`, changing the
callers namespace and the need to `cat` the file.

Signed-off-by: Michael Crosby <crosbymichael@gmail.com>