Do not require ipamV6Config if IPv6 is enabled for network
This brings IPv6 in line with IPv4 behaviour which does not require the
presence of an ipamV6Config on the network in order to trigger
allocating an address.

Since the enableIPv6 flag is an opt-in, this change does not alter the
default behaviour. The user has clearly indicated intent by explicitly
enabling the flag.

Signed-off-by: Thomas Graf <tgraf@suug.ch>