Split common interfaces into separate package
This decouples code that needs to import these interfaces with the rest
of the libnetwork codebase.

I noticed this originally working on Docker's container object and
seeing:

1. It takes a *really* long time to load, partially due to libnetwork
2. lots of platform dependent code from libnetwork makes the container
object not possible to compile except on target platforms, making the
test cycle a bit more difficult... this is even though the container
object is only dealing with these interfaces/common types.

Before this PR, the depdency tree just to use any of these interfaces
means importing serf, memberlist, protobuf, boltdb, consul/zk/etcd libs,
etc.

One thing this split is highlighting is that the interfaces for
`Sandbox`, and `Endpoint`, are very much tied to a specific
implementation.
This would be a nice future work to decouple further.

Signed-off-by: Brian Goff <cpuguy83@gmail.com>