Fix Google DNS fallback issue when `--dns-search/--dns-opt` is set
This fix tries to address the issue raised in:
https://github.com/docker/docker/issues/29815

where the Google DNS fallback does not work if either `--dns-search`
or `--dns-opt` is set.

The reason for the issue is that, in case either `--dns-search`
or `--dns-opt` is set, `resolvconf.FilterResolvDNS()` will not be
invoked to replace `localhost/127.*`.

This fix made changes in two places:
1. In `setupDNS()`, as long as `--dns` is not set, (`len(sb.config.dnsList) == 0`),
   `resolvconf.FilterResolvDNS()` will alwasy be invoked so that
   localhost could be replaced by Google DNS fallback.
2. In `updateDNS()`, as long as `--dns` is not set, (`len(sb.config.dnsList) == 0`),
   update will be processed.

The reason for (2) is that, at the time of `setupDNS()`, `ipv6Enabled` was still
unknown and `resolvconf.FilterResolvDNS()` always pass a `true` value. We need
`updateDNS()` to continue here to have the correct `ipv6Enabled` enabled and
adjust as needed.

Signed-off-by: Yong Tang <yong.tang.github@outlook.com>