control-plane hardning: cleanup local state on peer leaving a network
When the last task(s) are brought down on a node, network leave event is
also sent. network leave could potentially race with the task delete
events and if happens to complete first the task delete events won't be
sent out resulting in stale entries in the remote nodes. Clean up all
the state from a node for a network when a network leave event is
received from a node.

Signed-off-by: Santhosh Manohar <santhosh@docker.com>