Clean up kernel entries properly on task reschedules.
When a task gets scheduled on a node, its IP could have been used by
a task on a remote node that just went down. The add event could be
received before the delete the older task that went down. In this case
the kernel entry cleanup should be done before we update our state
based on the new add event. Otherwise the cleanup won't happen correctly
when the old delete event does arrive.

Signed-off-by: Santhosh Manohar <santhosh@docker.com>