Code cleanup using context
The debugs using the caller of method has been extremely useful
in faster debugging and understanding of flow. This commit uses context
to pass in the caller information instead of explicit method names.
The ideal way would have been to use the runtime package Caller function
to get the callers from the function stack instead of passing the method
explicitly. However since the caller tree need not be symmetric it would
make it ugly to retrieve the interested caller. Hencing moving the method
name passing to context which would make it cleaner and more readable.

Signed-off-by: Abhinandan Prativadi <abhi@docker.com>