Add Epoch ID for each instance of the network in the gossip
In swarm mode overlay networks are instantiated on a node when there is
a task attached to it and deleted when the last task goes away. When
there tasks come up and get deleted rapidly on a network this can lead
to whole bunch of network create/delete along with the associated
endpoint events for service-discovery and overlay. These endpoint events
can go out of order and can associated with wrong instance of the
network thus leading to incorrect SD or overlay programming.

This patch adds an epoch ID every time a network is added to the DB. Its
done at the network DB level so that its transparent to the clients and
the logic associated with filtering events based on the epcoh ID needn't
be exposed to multiple clients.

Signed-off-by: Santhosh Manohar <santhosh@docker.com>