Fix for crash on WaitGroup
panic: sync: WaitGroup is reused before previous Wait has returned

The panic is due to an EventNotify that arrives in conjunction with
a peerAdd or peerDelete. Wait is followed by an Add

The current WaitGroup logic does not add any further protection.
The pMap mutex should be enough

The update operations have to happen inside the pMap lock to be safe,
else there is a race between the update operations and new peerAdd and peerDelete

Signed-off-by: Flavio Crisciani <flavio.crisciani@docker.com>