Fixes race in overlay serf join.
Previously, multiple calls to nodeJoin could race the sync.Once which is typically called from
the new node to join it to the existing Serf pool. If the first peer encountered had failed,
serfJoin would fail, but the remaining (potentially operational) nodes may not be joined if
nodeJoin for these nodes had reached the sync.Once before the call to nodeJoin for the failed
peer had reached joinOnce = sync.Once().

Signed-off-by: Jamie Garside <jgarside@uk.ibm.com>