Simplifying 'possible transient state' log messages
When the number of entries in the endpoint set is large, the `setStr`
value will cause these log messages to be particularly long. In my
scenario I had ~1500-2000 entries, which were causing 100kB+ messages to
be logged.

This removes `setStr` from the message, as well as `b`, so that I can
also remove the call to `s.printIPToEndpoint`. I'm not sure how valuable
it is to know that the set is present, since we already know (from the
`ok` variable) that the mapping is present.

Signed-off-by: Dave Henderson <dhenderson@gmail.com>