Add comments to the resolv.conf file.
It can be confusing when you use the --dns x.x.x.x option on a container, to find it has not been added to resolv.conf
This is true for example, when you use the Macvlan network driver.  --dns is added to the Docker DNS server, and not resolv.conf.

I would suggest we see something like:

# --dns x.x.x.x has been added to Dockers internal DNS server
namserver 127.0.0.11
options ndots:0

or

# File set by Docker for the internal DNS server
namserver 127.0.0.11
options ndots:0

Signed-off-by: Stuart <stuart@logicethos.com>