NetworkDB node management race
The previous logic was not properly handling the case of a node
that was failing and coming back in short period of time.
The issue was in the handling of the network messages.
When a node joins, it sync state with other nodes, these are passing
the whole list of nodes that, at best of their knowledge, are part
of a network. At this point if the node receives that node A is part
of the network it saves it before having received the notification
that node A is actually alive (coming from memberlist).
If node A failed the source node will receive the notification
while the new joined node won't because memberlist never advertise
node A as available. In this case the new node will never purge
node A from its state but also worse, will accept any table notification
where node A is the owner and so will end up in a out of sync state
with the rest of the cluster.

Signed-off-by: Flavio Crisciani <flavio.crisciani@docker.com>
(cherry picked from commit 4702ccb)