resolver: less debug
Observed lots of debug messages from resolver (and associated high CPU
usage because of lots of logging) when debug logging is turned on. Here
is some data from a real system:

> $ journalctl -u docker --since="2019-10-21 21:00:00"  --until="2019-10-21 22:00:00"  | wc -l
> 188621
> $ journalctl -u docker --since="2019-10-21 21:00:00"  --until="2019-10-21 22:00:00"  | grep -E 'Name To resolve: |\[resolver\] ' | wc -l
> 186319

So, it was about 200000 lines logger for just one hour, and about 99%
of that are from resolver. While this might be the peculiarity of
a particular setup, the number of such messages still seem way too
excessive.

Remove the ones that are not errors. In case those are needed, one have
to recompile this package with debug_log build tag.

Signed-off-by: Kir Kolyshkin <kolyshkin@gmail.com>