update maintainers
The current list of maintainers is quite outdated, and we plan to
integrate libnetwork into the main Moby repository.

Before doing this, we may still need to review and merge fixes that
are pending in this repository (as well as maintain a branch for
the 20.10 release until it reaches EOL).

This patch adds some of the (currently) active maintainers of the
moby project to the maintainers list.

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>