portmapper: do not run dummy proxy when requesting a specific port
When a specific port is requested, we do not start a dummy proxy to
mark the port not available anymore. Doing so trigger numerous
problems, notably with UDP.

 - moby/moby#28589
 - moby/moby#8795
 - #2423

This change requires altering tests. It also becomes possible to map
to a port already used by the host, shadowing it. There is also the
possibility of using a port in the dynamic range which is already in
use by the host (but this case was not handled gracefully either,
making the whole allocation fails instead of trying the next port).

Alternatively, the change could be done only for UDP where the problem
is more apparent. Or it could be configurable.

Signed-off-by: Vincent Bernat <vincent@bernat.ch>